+++ allink_core/allink_config/models.py
@@ -216,17 +216,6 @@ class AllinkConfig(SingletonModel):
 from cms.extensions import TitleExtension
 from cms.extensions.extension_pool import extension_pool
 
-from cms.plugin_pool import plugin_pool
-from djangocms_file.cms_plugins import FilePlugin, FolderPlugin
-from djangocms_picture.cms_plugins import PicturePlugin
-
-plugin_pool.unregister_plugin(PicturePlugin)
-
-# Unregister existing Plugins
-
-plugin_pool.unregister_plugin(FilePlugin)
-plugin_pool.unregister_plugin(FolderPlugin)
-
 
 # Page Extensions
 class AllinkMetaTagExtension(AllinkMetaTagFieldsModel, TitleExtension):